src/python_analytics/requirements.txt
pandas==2.2.1
numpy==1.26.4
scikit-learn
matplotlib==3.8.3
seaborn==0.13.2
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
isort==5.13.2
mypy==1.8.0
pyarrow==15.0.0
python-snappy==0.6.1 